+++ docs/changes.md
@@ -9,6 +9,7 @@
 - Added 'AutoUpdate' property to camera, for better control over when it updates.
 - Added basic renderer with material override, so that animators could be used on sprites as well.
 - Added emissive color to override material.
+- Fixed sprite sheet default constructor.
 
 ## 0.1.0.7